+++ generators/app/questions/application.js
@@ -16,43 +16,87 @@ module.exports = [{
     message: 'Type your private key for JSON Web Token',
     default: crypto.randomBytes(32).toString('hex')
 }, {
+    type: 'confirm',
+    name: 'application:passport-facebook-enabled',
+    message: 'Do you need Facebook authorization?',
+    default: false
+}, {
     type: 'input',
-    name: 'application:facebook-client-id',
+    name: 'application:passport-facebook-client-id',
     message: 'Type your Facebook Client ID',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-facebook-enabled'];
+    }
 }, {
     type: 'input',
-    name: 'application:facebook-client-secret',
+    name: 'application:passport-facebook-client-secret',
     message: 'Type your Facebook Client Secret',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-facebook-enabled'];
+    }
+}, {
+    type: 'confirm',
+    name: 'application:passport-twitter-enabled',
+    message: 'Do you need Twitter authorization?',
+    default: false
 }, {
     type: 'input',
-    name: 'application:twitter-client-id',
+    name: 'application:passport-twitter-client-id',
     message: 'Type your Twitter Client ID',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-twitter-enabled'];
+    }
 }, {
     type: 'input',
-    name: 'application:twitter-client-secret',
+    name: 'application:passport-twitter-client-secret',
     message: 'Type your Twitter Client Secret',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-twitter-enabled'];
+    }
+}, {
+    type: 'confirm',
+    name: 'application:passport-yahoo-enabled',
+    message: 'Do you need Yahoo authorization?',
+    default: false
 }, {
     type: 'input',
-    name: 'application:yahoo-client-id',
+    name: 'application:passport-yahoo-client-id',
     message: 'Type your Yahoo Client ID',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-yahoo-enabled'];
+    }
 }, {
     type: 'input',
-    name: 'application:yahoo-client-secret',
+    name: 'application:passport-yahoo-client-secret',
     message: 'Type your Yahoo Client Secret',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-yahoo-enabled'];
+    }
+}, {
+    type: 'confirm',
+    name: 'application:passport-google-enabled',
+    message: 'Do you need Google Plus authorization?',
+    default: false
 }, {
     type: 'input',
-    name: 'application:google-client-id',
+    name: 'application:passport-google-client-id',
     message: 'Type your Google Client ID',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-google-enabled'];
+    }
 }, {
     type: 'input',
-    name: 'application:google-client-secret',
+    name: 'application:passport-google-client-secret',
     message: 'Type your Google Client Secret',
-    default: '-'
+    default: '-',
+    when: function (answers) {
+        return answers['application:passport-google-enabled'];
+    }
 }];