+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## Edge version
 
 - Improvement: Add Google Plus authorization;
-- Improvement: Add test coverage tool;
 - Improvement: Add Memory, Disk, SQLServer and OrientDB connections;
 - Improvement: Add model connection by default to Disk in development environment;
+- Improvement: Add questions kinda "Do you need Facebook authorization?";
+- Improvement: Add test coverage tool;
 
 ## Version 0.6.2